requirements.txt
higher==0.2.1
jax==0.2.19
jaxlib==0.1.70+cuda102
kornia==0.7.0
matplotlib==3.3.2
neural-tangents==0.3.8
numpy==1.24.3
Pillow==9.3.0
quadprog==0.1.11
scikit-image==0.21.0
scikit-learn==0.23.2
scipy==1.10.1
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tqdm==4.36.1
Wand==0.6.13
six==1.16.0
onedrivedownloader==1.1.3
setproctitle==1.3.3
